----------------------------------------------------------------------  Enter Log.  Lines beginning with `CVS:' are removed automatically
 Committing in .

 Modified Files:
 	config/configure.status
 	src/library/algorithms/Gauss/lin_dom_gauss.C
 	src/library/archetypes/element/element_archetype.h
 	src/library/archetypes/field/field_abstract.h
 	src/library/archetypes/field/field_archetype.h
 	src/library/archetypes/field/field_envelope.h
 	src/library/archetypes/random/randiter_abstract.h
 	src/library/archetypes/random/randiter_archetype.h
 	src/library/archetypes/random/randiter_envelope.h
 	src/library/objects/BlackBox/lin_erase_bb.h
 	src/tools/SparseVector/lin_pair.h
 ----------------------------------------------------------------------

+++ trunk/linbox-old/src/library/archetypes/random/randiter_abstract.h
@@ -4,6 +4,7 @@
 
 #ifndef _RANDITER_ABSTRACT_
 #define _RANDITER_ABSTRACT_
+#include "LinBox/integer.h"
 
 #include <iostream>
 
@@ -46,7 +47,7 @@ namespace LinBox
      *             (default = -1)
      */
     virtual RandIter_abstract* construct(const Field_abstract& F, 
-					 size_t size = 0, 
+					 integer size = 0, 
 					 size_t seed = size_t(-1)) const = 0;
 
     /** Virtual copy constructor.
